crypto-trading-system/pyproject.toml
[project]
name = "crypto-trading-system"
version = "0.1.0"
description = "Phase 1 MVP multi-agent crypto paper trading system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "apscheduler>=3.10.4",
  "httpx>=0.27.0",
  "pydantic>=2.6.0",
  "pydantic-settings>=2.2.0",
  "pyyaml>=6.0.1",
  "sqlalchemy>=2.0.28",
  "psycopg2-binary>=2.9.9",
  "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
